+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### [1.3.0](../../releases/tag/v1.3.0) - 2026-09-19
 - Drop support for Python3.9
 - Migrate from poetry to uv
+- Requires anyio
 
 ## 1.2